+++ webapp/tests/unit/user/user-has-local-password-schema.test.js
@@ -5,8 +5,8 @@
  *                   applyDefaults() intentionally leaves it unset for new local users (absent
  *                   field reads as true - no migration/backfill needed, see model/user.js)
  * @file            webapp/tests/unit/user/user-has-local-password-schema.test.js
- * @version         1.7.5
- * @release         2026-07-30
+ * @version         1.7.6
+ * @release         2026-07-31
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025-2026 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/