+++ webapp/tests/unit/utils/crypto-secrets.test.js
@@ -4,8 +4,8 @@
  * @description     Round-trip, salt-namespacing, and tamper-detection tests for
  *                   encryptSecret()/decryptSecret()
  * @file            webapp/tests/unit/utils/crypto-secrets.test.js
- * @version         1.7.5
- * @release         2026-07-30
+ * @version         1.7.6
+ * @release         2026-07-31
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025-2026 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/