+++ webapp/utils/theme-manager.js
@@ -3,8 +3,8 @@
  * @tagline         Theme discovery and schema extension utility
  * @description     Discovers themes from framework, plugins, and site; extends UserModel theme enum
  * @file            webapp/utils/theme-manager.js
- * @version         1.7.5
- * @release         2026-07-30
+ * @version         1.7.6
+ * @release         2026-07-31
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/